+++ mirrord/layer/src/lib.rs
@@ -581,9 +581,9 @@ pub(crate) unsafe extern "C" fn close_detour(fd: c_int) -> c_int {
         .get()
         .expect("Should be set during initialization!");
 
-    if SOCKETS.lock().unwrap().remove(&fd).is_none() && *enabled_file_ops
-        && let Some(remote_fd) = OPEN_FILES.lock().unwrap().remove(&fd) {
-        let close_file_result = file::ops::close(remote_fd);
+    if SOCKETS.lock().unwrap().remove(&fd).is_some() && *enabled_file_ops
+        && let Some(local_file) = OPEN_FILES.lock().unwrap().remove(&fd) {
+        let close_file_result = file::ops::close(local_file.resource);
 
         close_file_result
             .map_err(|fail| {